requirements.txt
-i https://pypi.org/simple

flask==3.0.3
sqlalchemy==2.0.29
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-swagger==0.2.14
python-dotenv==1.0.1
flask-cors==4.0.1
gunicorn==22.0.0
cloudinary==1.39.0
flask-admin==1.6.1
typing-extensions==4.11.0
flask-jwt-extended==4.6.0
wtforms==3.1.2
psycopg2-binary==2.9.9
alembic==1.13.1
certifi==2024.2.2
click==8.1.7
itsdangerous==2.2.0
jinja2==3.1.4
mako==1.3.3
markupsafe==2.1.5
python-dateutil==2.9.0.post0
python-editor==1.0.4
pyyaml==6.0.1
six==1.16.0
urllib3==2.2.1
werkzeug==3.0.2
flask-mail==0.9.1
python-dotenv==1.0.1
bcrypt==4.1.2
cloudinary == 2.6.1

